requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
django-storages
boto3
psycopg2-binary
gunicorn
